gave up on squire
I wanted to have more reactive art for this story. but it already has some, and I'm not sure it needs any more.

Showing the squire ageing might remove some of the impact of the dragon revealing to the audience that the squire is an old man.

Plus, I don't really know exactly how I want to style the people. :P