Consolidated header file
All the pages have the same interactive banner, so I figured I should just put that header into one file so I can change it in them all as needed.

Unfortunately all the headers have the map, which needs to be slightly different depending on which page is the user is at.

The solution is fairly straightforward (store the current page in a variable instead of in the 'onclick()' code, like I was doing, because I'm an idiot) so obviously I didn't figure it out until I'd exhausted every other possibility after several hours (because I'm an idiot)

Should help in the future tho.