transitions for The_wizard_girl
That took a while, but I have everything dialed in how I want it, I think.

Except the audio, which means I'll need to redo it, and redo all the things that were synced to it. which was all the timings I just did. So that's great.

Once that's done though, I'll be finished with the page I think. :D